refactor: move Result at top
