+++ src/type.ts
@@ -30,3 +30,10 @@ export type FailedLintTableRecord = {
   test: string
   message: string
 }
+
+export enum Result {
+  Passed = 'passed',
+  Failed = 'failed',
+  Skipped = 'skipped',
+  Unknown = 'unknown'
+}
